+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 ## 1.0.1
 - Add example
 - Fix formatting
+
+## 1.0.2
+- Renamed showAllButton to changeExpansionButton
+- Added the `isLimited` parameter, which controls the visibility of the changeExpansionButton
+- Updated readme and test coverage